+++ main.py
@@ -6,7 +6,6 @@ import pandas as pd
 
 import library.UI as UI
 import library.analysis as analysis
-import library.processes as processes
 
 logging.basicConfig(level=logging.INFO,
                     format='%(name)s - %(levelname)s - %(message)s')
@@ -29,16 +28,16 @@ if __name__ == "__main__":
 
     logging.info(f"File path: {args.file}")
 
-    print(f"1. Loading file...")
-    movements = processes.load_movements(args.file)
+    # print(f"1. Loading file...")
+    # movements = processes.load_movements(args.file)
     
-    print(f"2. Recognise movements and wealth columns...")
-    mov_col, wea_col = analysis.detect_money_cols(movements)
-    print(f"\tMovements column: {mov_col}")
-    print(f"\tWealth column: {wea_col}")
+    # print(f"2. Recognise movements and wealth columns...")
+    # mov_col, wea_col = analysis.detect_money_cols(movements)
+    # print(f"\tMovements column: {mov_col}")
+    # print(f"\tWealth column: {wea_col}")
     
-    print(f"3. Recognise dates columns...")
-    date_cols = analysis.detect_dates_cols(movements)
-    print(f"\tColumns: {date_cols}")
+    # print(f"3. Recognise dates columns...")
+    # date_cols = analysis.detect_dates_cols(movements)
+    # print(f"\tColumns: {date_cols}")
 
     UI.menu()